+++ man-roxygen/param_sensors.R
@@ -1,2 +1,1 @@
-#' @param localities names of localities; if NULL then all (default NULL)
 #' @param sensors names of sensors; if NULL then all (default NULL) see `names(mc_data_sensors)`